requirements.txt
Cython>=0.20
python-rocksdb
